Rename vote mocks to voter
This name change reflects the name of the query that the mock pertains toi
