+++ src/clients/api/queries/getVoters/formatToVoters.ts
