+++ src/pages/Proposal/index.stories.tsx
@@ -3,7 +3,7 @@ import noop from 'noop-ts';
 import { ComponentMeta } from '@storybook/react';
 import { withRouter } from 'stories/decorators';
 import proposals from '__mocks__/models/proposals';
-import vote from '__mocks__/models/vote';
+import voters from '__mocks__/models/voters';
 import { ProposalUi } from '.';
 
 export default {
@@ -15,9 +15,9 @@ export default {
 export const Default = () => (
   <ProposalUi
     proposal={proposals[0]}
-    forVoters={vote}
-    againstVoters={vote}
-    abstainVoters={vote}
+    forVoters={voters}
+    againstVoters={voters}
+    abstainVoters={voters}
     votingEnabled
     readableVoteWeight="123390000"
     vote={noop}
